FormCondition, EqualConditioin in its own file
